INC (40,41,42,43,44,45,46,47; real,pe) + implemented lazy eflags logic

+++ lib86cpu/lib86cpu.h
@@ -120,7 +120,7 @@ struct regs_layout_t {
 
 struct lazy_eflags_t {
 	uint32_t result;
-	uint32_t auxbits;
+	uint8_t auxbits;
 };
 
 struct cpu_t {
